Remove unused AggregationWriter
